chore: Solve some circular dependencies
