+++ src/components/DeployButton/DeployButton.types.ts
@@ -1,7 +1,7 @@
 import { Dispatch } from 'redux'
 import { openModal, OpenModalAction } from 'modules/modal/actions'
 import { DeploymentStatus } from 'modules/deployment/types'
-import { ModelMetrics } from 'modules/scene/types'
+import { ModelMetrics } from 'modules/models/types'
 import { Project } from 'modules/project/types'
 
 export type Props = {
