+++ src/components/TopBar/TopBar.types.ts
@@ -11,7 +11,7 @@ import {
 } from 'modules/editor/actions'
 import { resetItem, duplicateItem, deleteItem, ResetItemAction, DuplicateItemAction, DeleteItemAction } from 'modules/scene/actions'
 import { openModal, OpenModalAction } from 'modules/modal/actions'
-import { ModelMetrics } from 'modules/scene/types'
+import { ModelMetrics } from 'modules/models/types'
 import { Project } from 'modules/project/types'
 import { Gizmo } from 'modules/editor/types'
 import { PoolGroup } from 'modules/poolGroup/types'
