+++ src/modules/common/sagas.ts
@@ -33,10 +33,11 @@ import { uiSaga } from 'modules/ui/sagas'
 import { walletSaga } from 'modules/wallet/sagas'
 
 import { PEER_URL } from 'lib/api/peer'
+import { BuilderAPI } from 'lib/api/builder'
 
 const profileSaga = createProfileSaga({ peerUrl: PEER_URL })
 
-export function* rootSaga() {
+export function* rootSaga(builderAPI: BuilderAPI) {
   yield all([
     analyticsSaga(),
     assetPackSaga(),
