+++ src/modules/editor/utils.ts
@@ -4,7 +4,7 @@ import { Project } from 'modules/project/types'
 import { getSceneDefinition } from 'modules/project/export'
 import { getContentsStorageUrl } from 'lib/api/builder'
 import { capitalize } from 'lib/text'
-import { Vector3 } from 'modules/common/types'
+import { Vector3 } from 'modules/models/types'
 import { Scene, EntityDefinition, ComponentDefinition, ComponentType } from 'modules/scene/types'
 import { TRANSPARENT_PIXEL } from 'lib/getModelData'
 import { getMetrics } from 'components/AssetImporter/utils'
