+++ src/modules/identity/utils.ts
@@ -1,14 +1,9 @@
 import { select, put, race, take, delay } from 'redux-saga/effects'
-import { AuthIdentity } from 'dcl-crypto'
 import { getCurrentIdentity, isLoggedIn } from 'modules/identity/selectors'
 import { openModal } from 'modules/modal/actions'
 import { LOGIN_SUCCESS, LOGIN_FAILURE } from './actions'
 
 export const ONE_MONTH_IN_MINUTES = 31 * 24 * 60
-
-export function isValid(identity?: AuthIdentity | null) {
-  return !!identity && Date.now() < +new Date(identity.expiration)
-}
 
 // Helper that always yields a valid identity
 export function* getIdentity(): IterableIterator<any> {
