+++ src/modules/item/types.ts
@@ -1,4 +1,4 @@
-import { ModelMetrics } from 'modules/scene/types'
+import { ModelMetrics } from 'modules/models/types'
 
 export enum ItemType {
   WEARABLE = 'wearable'
