+++ src/modules/media/sagas.ts
@@ -2,7 +2,7 @@ import { select, delay, put, call, takeLatest } from 'redux-saga/effects'
 import { Omit } from 'decentraland-dapps/dist/lib/types'
 import { getCurrentProject } from 'modules/project/selectors'
 import { dataURLToBlob } from 'modules/media/utils'
-import { PARCEL_SIZE } from 'modules/project/utils'
+import { PARCEL_SIZE } from 'modules/project/constants'
 import { Project } from 'modules/project/types'
 import { EditorWindow } from 'components/Preview/Preview.types'
 import { RECORD_MEDIA_REQUEST, recordMediaProgress, recordMediaSuccess } from './actions'
