+++ src/modules/project/sagas.ts
@@ -211,10 +211,8 @@ function* handleExportProject(action: ExportProjectRequestAction) {
     zip.file(filename, files[filename])
   }
 
-  yield call(async () => {
-    const artifact = await zip.generateAsync<'blob'>({ type: 'blob' })
-    saveAs(artifact, `${sanitizedName}.zip`)
-  })
+  const artifact = yield call(zip.generateAsync, { type: 'blob' })
+  saveAs(artifact, `${sanitizedName}.zip`)
 
   yield put(closeModal('ExportModal'))
   yield put(exportProjectSuccess())
