+++ src/modules/scene/sagas.ts
@@ -38,7 +38,7 @@ import { ComponentType, Scene, ComponentDefinition, ShapeComponent, AnyComponent
 import { getSelectedEntityIds, isReady } from 'modules/editor/selectors'
 import { setSelectedEntities, SET_EDITOR_READY } from 'modules/editor/actions'
 import { getCurrentBounds, getData as getProjects } from 'modules/project/selectors'
-import { PARCEL_SIZE } from 'modules/project/utils'
+import { PARCEL_SIZE } from 'modules/project/constants'
 import { EditorWindow } from 'components/Preview/Preview.types'
 import { COLLECTIBLE_ASSET_PACK_ID } from 'modules/ui/sidebar/utils'
 import {
